# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"_36_Enums.RequestStatus": {
				"type": "string",
				"enum": [
					"PENDING",
					"ASSIGNED",
					"IN_PROGRESS",
					"COMPLETED",
					"CANCELLED"
				]
			},
			"CreateWebDigitalRequest": {
				"properties": {
					"customerName": {
						"type": "string"
					},
					"customerPhone": {
						"type": "string"
					},
					"customerEmail": {
						"type": "string"
					},
					"service": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"documentUrl": {
						"type": "string"
					},
					"preferredDate": {
						"type": "string"
					},
					"tasker": {
						"type": "string"
					}
				},
				"required": [
					"customerName",
					"customerPhone",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateWebDigitalRequest": {
				"properties": {
					"customerName": {
						"type": "string"
					},
					"customerPhone": {
						"type": "string"
					},
					"customerEmail": {
						"type": "string"
					},
					"service": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"documentUrl": {
						"type": "string"
					},
					"preferredDate": {
						"type": "string"
					},
					"tasker": {
						"type": "string"
					}
				},
				"required": [
					"customerName",
					"customerPhone",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "rise-motive-be",
		"version": "1.0.0",
		"description": "Rise Motive Backend API",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/web-digital": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "number",
											"format": "double"
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"tasker": {
											"type": "string"
										},
										"status": {
											"$ref": "#/components/schemas/_36_Enums.RequestStatus"
										},
										"preferredDate": {
											"type": "string"
										},
										"documentUrl": {
											"type": "string"
										},
										"description": {
											"type": "string"
										},
										"service": {
											"type": "string"
										},
										"customerEmail": {
											"type": "string"
										},
										"customerPhone": {
											"type": "string"
										},
										"customerName": {
											"type": "string"
										},
										"trackingCode": {
											"type": "string"
										}
									},
									"required": [
										"id",
										"updatedAt",
										"createdAt",
										"tasker",
										"status",
										"preferredDate",
										"documentUrl",
										"description",
										"service",
										"customerEmail",
										"customerPhone",
										"customerName",
										"trackingCode"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Web & Digital"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateWebDigitalRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"items": {
											"items": {
												"properties": {
													"id": {
														"type": "number",
														"format": "double"
													},
													"updatedAt": {
														"type": "string",
														"format": "date-time"
													},
													"createdAt": {
														"type": "string",
														"format": "date-time"
													},
													"tasker": {
														"type": "string"
													},
													"status": {
														"$ref": "#/components/schemas/_36_Enums.RequestStatus"
													},
													"preferredDate": {
														"type": "string"
													},
													"documentUrl": {
														"type": "string"
													},
													"description": {
														"type": "string"
													},
													"service": {
														"type": "string"
													},
													"customerEmail": {
														"type": "string"
													},
													"customerPhone": {
														"type": "string"
													},
													"customerName": {
														"type": "string"
													},
													"trackingCode": {
														"type": "string"
													}
												},
												"required": [
													"id",
													"updatedAt",
													"createdAt",
													"tasker",
													"status",
													"preferredDate",
													"documentUrl",
													"description",
													"service",
													"customerEmail",
													"customerPhone",
													"customerName",
													"trackingCode"
												],
												"type": "object"
											},
											"type": "array"
										},
										"total": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"items",
										"total"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Web & Digital"
				],
				"security": [],
				"parameters": []
			}
		},
		"/web-digital/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "number",
											"format": "double"
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"tasker": {
											"type": "string"
										},
										"status": {
											"$ref": "#/components/schemas/_36_Enums.RequestStatus"
										},
										"preferredDate": {
											"type": "string"
										},
										"documentUrl": {
											"type": "string"
										},
										"description": {
											"type": "string"
										},
										"service": {
											"type": "string"
										},
										"customerEmail": {
											"type": "string"
										},
										"customerPhone": {
											"type": "string"
										},
										"customerName": {
											"type": "string"
										},
										"trackingCode": {
											"type": "string"
										}
									},
									"required": [
										"id",
										"updatedAt",
										"createdAt",
										"tasker",
										"status",
										"preferredDate",
										"documentUrl",
										"description",
										"service",
										"customerEmail",
										"customerPhone",
										"customerName",
										"trackingCode"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Web & Digital"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "number",
											"format": "double"
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"tasker": {
											"type": "string"
										},
										"status": {
											"$ref": "#/components/schemas/_36_Enums.RequestStatus"
										},
										"preferredDate": {
											"type": "string"
										},
										"documentUrl": {
											"type": "string"
										},
										"description": {
											"type": "string"
										},
										"service": {
											"type": "string"
										},
										"customerEmail": {
											"type": "string"
										},
										"customerPhone": {
											"type": "string"
										},
										"customerName": {
											"type": "string"
										},
										"trackingCode": {
											"type": "string"
										}
									},
									"required": [
										"id",
										"updatedAt",
										"createdAt",
										"tasker",
										"status",
										"preferredDate",
										"documentUrl",
										"description",
										"service",
										"customerEmail",
										"customerPhone",
										"customerName",
										"trackingCode"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Web & Digital"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateWebDigitalRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "number",
											"format": "double"
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"tasker": {
											"type": "string"
										},
										"status": {
											"$ref": "#/components/schemas/_36_Enums.RequestStatus"
										},
										"preferredDate": {
											"type": "string"
										},
										"documentUrl": {
											"type": "string"
										},
										"description": {
											"type": "string"
										},
										"service": {
											"type": "string"
										},
										"customerEmail": {
											"type": "string"
										},
										"customerPhone": {
											"type": "string"
										},
										"customerName": {
											"type": "string"
										},
										"trackingCode": {
											"type": "string"
										}
									},
									"required": [
										"id",
										"updatedAt",
										"createdAt",
										"tasker",
										"status",
										"preferredDate",
										"documentUrl",
										"description",
										"service",
										"customerEmail",
										"customerPhone",
										"customerName",
										"trackingCode"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Web & Digital"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}